add .connect files
