+++ src/components/app/app.connect.jsx
@@ -1,0 +1,8 @@
+import {connect} from "react-redux";
+import App from "@components/app/app";
+
+const mapStateToProps = (state) => {
+  return state;
+};
+
+export default connect(mapStateToProps)(App);
