+++ src/components/app/app.test.js
@@ -1,4 +1,4 @@
-import {App} from "./app";
+import App from "./app";
 import {Provider} from "react-redux";
 import configureStore from "redux-mock-store";
