+++ src/components/genre-list/genre-list.connect.jsx
@@ -1,0 +1,17 @@
+import {ActionCreator} from "../../reducer";
+import {connect} from "react-redux";
+import GenreList from "@components/genre-list/genre-list";
+
+const mapStateToProps = (state) => ({
+  genres: state.genres,
+  activeGenre: state.activeGenre
+});
+
+const mapDispatchToProps = (dispatch) => ({
+  onGenreChange(genre) {
+    dispatch(ActionCreator.setActiveGenre(genre));
+    dispatch(ActionCreator.getFilteredMovies());
+  },
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(GenreList);
