+++ src/index.jsx
@@ -1,7 +1,7 @@
 import ReactDOM from 'react-dom';
 import {createStore} from "redux";
 import {Provider} from "react-redux";
-import App from '@components/app/app';
+import App from '@components/app/app.connect';
 import {MainMovie} from './constants';
 import {reducer} from "./reducer.js";
